Refactor snapshot into a DTO
