+++ restic/dto/snapshot.go
@@ -1,0 +1,17 @@
+package dto
+
+import "time"
+
+// Snapshot models a restic a single snapshot from the
+// snapshots --json subcommand.
+type Snapshot struct {
+	ID       string    `json:"id"`
+	Time     time.Time `json:"time"`
+	Tree     string    `json:"tree"`
+	Paths    []string  `json:"paths"`
+	Hostname string    `json:"hostname"`
+	Username string    `json:"username"`
+	UID      int       `json:"uid"`
+	Gid      int       `json:"gid"`
+	Tags     []string  `json:"tags"`
+}
